Remove README link from all setplot.py files.
Also move some setplot.py files that I missed previously
into the accompanying script.
