+++ examples/advection_2d_annulus/advection_annulus.py
@@ -275,6 +275,63 @@ def setup(use_petsc=False,outdir='./_output',solver_type='classic'):
     return claw
 
 
+#--------------------------
+def setplot(plotdata):
+#--------------------------
+    """ 
+    Specify what is to be plotted at each frame.
+    Input:  plotdata, an instance of visclaw.data.ClawPlotData.
+    Output: a modified version of plotdata.
+    """
+    from mapc2p import mapc2p
+    import numpy as np
+    from clawpack.visclaw import colormaps
+
+    plotdata.clearfigures()  # clear any old figures,axes,items data
+    
+    # Figure for pcolor plot
+    plotfigure = plotdata.new_plotfigure(name='q[0]', figno=0)
+
+    # Set up for axes in this figure:
+    plotaxes = plotfigure.new_plotaxes()
+    plotaxes.xlimits = 'auto'
+    plotaxes.ylimits = 'auto'
+    plotaxes.title = 'q[0]'
+    plotaxes.afteraxes = "pylab.axis('scaled')" 
+
+    # Set up for item on these axes:
+    plotitem = plotaxes.new_plotitem(plot_type='2d_pcolor')
+    plotitem.plot_var = 0
+    plotitem.pcolor_cmap = colormaps.red_yellow_blue
+    plotitem.pcolor_cmin = -1.
+    plotitem.pcolor_cmax = 1.
+    plotitem.add_colorbar = True
+    plotitem.MappedGrid = True
+    plotitem.mapc2p = mapc2p
+
+
+    # Figure for contour plot
+    plotfigure = plotdata.new_plotfigure(name='contour', figno=1)
+
+    # Set up for axes in this figure:
+    plotaxes = plotfigure.new_plotaxes()
+    plotaxes.xlimits = 'auto'
+    plotaxes.ylimits = 'auto'
+    plotaxes.title = 'q[0]'
+    plotaxes.scaled = True
+
+    # Set up for item on these axes:
+    plotitem = plotaxes.new_plotitem(plot_type='2d_contour')
+    plotitem.plot_var = 0
+    plotitem.contour_levels = np.linspace(-0.9, 0.9, 10)
+    plotitem.contour_colors = 'k'
+    plotitem.patchedges_show = 1
+    plotitem.MappedGrid = True
+    plotitem.mapc2p = mapc2p
+
+    return plotdata
+
+ 
 if __name__=="__main__":
     from clawpack.pyclaw.util import run_app_from_main
-    output = run_app_from_main(setup)
+    output = run_app_from_main(setup,setplot)
