--- examples/advection_2d_annulus/setplot.py
@@ -1,93 +1,0 @@
-
-""" 
-Set up the plot figures, axes, and items to be done for each frame.
-
-This module is imported by the plotting routines and then the
-function setplot is called to set the plot parameters.
-    
-""" 
-
-from mapc2p import mapc2p
-import numpy as np
-
-
-#--------------------------
-def setplot(plotdata):
-#--------------------------
-    
-    """ 
-    Specify what is to be plotted at each frame.
-    Input:  plotdata, an instance of visclaw.data.ClawPlotData.
-    Output: a modified version of plotdata.
-    
-    """
-
-
-    from clawpack.visclaw import colormaps
-
-    plotdata.clearfigures()  # clear any old figures,axes,items data
-    
-    
-    # Figure for pcolor plot
-    plotfigure = plotdata.new_plotfigure(name='q[0]', figno=0)
-
-    # Set up for axes in this figure:
-    plotaxes = plotfigure.new_plotaxes()
-    plotaxes.xlimits = 'auto'
-    plotaxes.ylimits = 'auto'
-    plotaxes.title = 'q[0]'
-    plotaxes.afteraxes = "pylab.axis('scaled')" 
-
-    # Set up for item on these axes:
-    plotitem = plotaxes.new_plotitem(plot_type='2d_pcolor')
-    plotitem.plot_var = 0
-    plotitem.pcolor_cmap = colormaps.red_yellow_blue
-    plotitem.pcolor_cmin = -1.
-    plotitem.pcolor_cmax = 1.
-    plotitem.add_colorbar = True
-    plotitem.MappedGrid = True
-    plotitem.mapc2p = mapc2p
-    plotitem.show = True       # show on plot?
-
-
-    # Figure for contour plot
-    plotfigure = plotdata.new_plotfigure(name='contour', figno=1)
-
-    # Set up for axes in this figure:
-    plotaxes = plotfigure.new_plotaxes()
-    plotaxes.xlimits = 'auto'
-    plotaxes.ylimits = 'auto'
-    plotaxes.title = 'q[0]'
-    plotaxes.scaled = True
-
-    # Set up for item on these axes:
-    plotitem = plotaxes.new_plotitem(plot_type='2d_contour')
-    plotitem.plot_var = 0
-    plotitem.contour_levels = np.linspace(-0.9, 0.9, 10)
-    plotitem.contour_colors = 'k'
-    plotitem.patchedges_show = 1
-    plotitem.MappedGrid = True
-    plotitem.mapc2p = mapc2p
-    plotitem.show = True       # show on plot?
-
-
-    # Parameters used only when creating html and/or latex hardcopy
-    # e.g., via visclaw.frametools.printframes:
-
-    plotdata.printfigs = True                # print figures
-    plotdata.print_format = 'png'            # file format
-    plotdata.print_framenos = 'all'          # list of frames to print
-    plotdata.print_fignos = 'all'            # list of figures to print
-    plotdata.html = True                     # create html files of plots?
-    plotdata.html_homelink = '../README.html'   # pointer for top of index
-    plotdata.latex = True                    # create latex file of plots?
-    plotdata.latex_figsperline = 2           # layout of plots
-    plotdata.latex_framesperline = 1         # layout of plots
-    plotdata.latex_makepdf = False           # also run pdflatex?
-
-    return plotdata
-
-    
-
-
-
